requirements.txt
# ─── API / orchestration ──────────────────────────────
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-dotenv==1.0.1
pydantic==2.9.2
langchain==0.3.4
langchain-core==0.3.12

# ─── Intent processor ────────────────────────────────
# Local pipeline (intent_pipeline.py): BART NLI + SBERT + SymSpell
torch>=2.2.0
transformers>=4.44.0
sentence-transformers==3.1.1
symspellpy>=6.7.7
# LLM pipeline (llm_intent_pipeline.py): Groq / Cerebras / OpenRouter / SambaNova
# all OpenAI-compatible — just need the openai SDK
openai>=1.40.0

# ─── Hybrid retrieval ─────────────────────────────────
rank-bm25==0.2.2
faiss-cpu==1.8.0.post1
numpy==1.26.4
pandas==2.2.3

# ─── Knowledge graph ──────────────────────────────────
neo4j==5.25.0

# ─── LLM (local via Ollama) ───────────────────────────
ollama==0.3.3

# ─── Watchers ─────────────────────────────────────────
watchdog==5.0.3
thefuzz==0.22.1
python-Levenshtein==0.26.0
